python-script/requirements.txt
gspread
oauth2client
requests
python-dotenv